+++ aerich/inspectdb/postgres.py
@@ -1,17 +1,20 @@
-from typing import List, Optional
+from __future__ import annotations
 
-from tortoise import BaseDBAsyncClient
+from typing import TYPE_CHECKING
 
-from aerich.inspectdb import Column, Inspect
+from aerich.inspectdb import Column, FieldMapDict, Inspect
+
+if TYPE_CHECKING:
+    from tortoise.backends.base_postgres.client import BasePostgresClient
 
 
 class InspectPostgres(Inspect):
-    def __init__(self, conn: BaseDBAsyncClient, tables: Optional[List[str]] = None):
+    def __init__(self, conn: "BasePostgresClient", tables: list[str] | None = None) -> None:
         super().__init__(conn, tables)
-        self.schema = self.conn.server_settings.get("schema") or "public"
+        self.schema = conn.server_settings.get("schema") or "public"
 
     @property
-    def field_map(self) -> dict:
+    def field_map(self) -> FieldMapDict:
         return {
             "int4": self.int_field,
             "int8": self.int_field,
@@ -33,12 +36,12 @@ class InspectPostgres(Inspect):
             "timestamp": self.datetime_field,
         }
 
-    async def get_all_tables(self) -> List[str]:
+    async def get_all_tables(self) -> list[str]:
         sql = "select TABLE_NAME from information_schema.TABLES where table_catalog=$1 and table_schema=$2"
         ret = await self.conn.execute_query_dict(sql, [self.database, self.schema])
         return list(map(lambda x: x["table_name"], ret))
 
-    async def get_columns(self, table: str) -> List[Column]:
+    async def get_columns(self, table: str) -> list[Column]:
         columns = []
         sql = f"""select c.column_name,
        col_description('public.{table}'::regclass, ordinal_position) as column_comment,
@@ -55,7 +58,7 @@ from information_schema.constraint_column_usage const
          right join information_schema.columns c using (column_name, table_catalog, table_schema, table_name)
 where c.table_catalog = $1
   and c.table_name = $2
-  and c.table_schema = $3"""
+  and c.table_schema = $3"""  # nosec:B608
         ret = await self.conn.execute_query_dict(sql, [self.database, table, self.schema])
         for row in ret:
             columns.append(