+++ tests/test_migrate.py
@@ -1,11 +1,15 @@
+from pathlib import Path
+from typing import List, cast
+
 import pytest
+import tortoise
 from pytest_mock import MockerFixture
 
 from aerich.ddl.mysql import MysqlDDL
 from aerich.ddl.postgres import PostgresDDL
 from aerich.ddl.sqlite import SqliteDDL
 from aerich.exceptions import NotSupportError
-from aerich.migrate import Migrate
+from aerich.migrate import MIGRATE_TEMPLATE, Migrate
 from aerich.utils import get_models_describe
 
 old_models_describe = {
@@ -100,6 +104,21 @@ old_models_describe = {
                 "constraints": {"ge": 1, "le": 2147483647},
                 "db_field_types": {"": "INT"},
             },
+            {
+                "name": "title",
+                "field_type": "CharField",
+                "db_column": "title",
+                "python_type": "str",
+                "generated": False,
+                "nullable": False,
+                "unique": True,
+                "indexed": True,
+                "default": None,
+                "description": None,
+                "docstring": None,
+                "constraints": {"max_length": 20},
+                "db_field_types": {"": "VARCHAR(20)"},
+            },
         ],
         "fk_fields": [
             {
@@ -169,6 +188,21 @@ old_models_describe = {
             "db_field_types": {"": "INT"},
         },
         "data_fields": [
+            {
+                "name": "name",
+                "field_type": "CharField",
+                "db_column": "name",
+                "python_type": "str",
+                "generated": False,
+                "nullable": False,
+                "unique": True,
+                "indexed": True,
+                "default": None,
+                "description": None,
+                "docstring": None,
+                "constraints": {"max_length": 100},
+                "db_field_types": {"": "VARCHAR(100)"},
+            },
             {
                 "name": "label",
                 "field_type": "CharField",
@@ -368,6 +402,21 @@ old_models_describe = {
                 "docstring": None,
                 "constraints": {"max_length": 50},
                 "db_field_types": {"": "VARCHAR(50)"},
+            },
+            {
+                "name": "uid",
+                "field_type": "IntField",
+                "db_column": "uuid",
+                "python_type": "int",
+                "generated": False,
+                "nullable": False,
+                "unique": True,
+                "indexed": True,
+                "default": None,
+                "description": None,
+                "docstring": None,
+                "constraints": {"ge": -2147483648, "le": 2147483647},
+                "db_field_types": {"": "INT"},
             },
             {
                 "name": "view_num",
@@ -773,6 +822,16 @@ old_models_describe = {
 }
 
 
+def should_add_user_id_column_type_alter_sql() -> bool:
+    if tortoise.__version__ < "0.21":
+        return False
+    # tortoise-orm>=0.21 changes IntField constraints
+    # from {"ge": 1, "le": 2147483647} to {"ge": -2147483648, "le": 2147483647}
+    data_fields = cast(List[dict], old_models_describe["models.Category"]["data_fields"])
+    user_id_constraints = data_fields[-1]["constraints"]
+    return tortoise.fields.data.IntField.constraints != user_id_constraints
+
+
 def test_migrate(mocker: MockerFixture):
     """
     models.py diff with old_models.py
@@ -783,28 +842,37 @@ def test_migrate(mocker: MockerFixture):
     - drop field: User.avatar
     - add index: Email.email
     - add many to many: Email.users
-    - remove unique: User.username
+    - remove unique: Category.title
+    - add unique: User.username
     - change column: length User.password
     - add unique_together: (name,type) of Product
+    - add one more many to many field: Product.users
+    - drop unique field: Config.name
     - alter default: Config.status
     - rename column: Product.image -> Product.pic
     """
-    mocker.patch("click.prompt", side_effect=(True,))
+    mocker.patch("asyncclick.prompt", side_effect=(True,))
 
     models_describe = get_models_describe("models")
     Migrate.app = "models"
     if isinstance(Migrate.ddl, SqliteDDL):
         with pytest.raises(NotSupportError):
             Migrate.diff_models(old_models_describe, models_describe)
+        Migrate.upgrade_operators.clear()
+        with pytest.raises(NotSupportError):
             Migrate.diff_models(models_describe, old_models_describe, False)
+        Migrate.downgrade_operators.clear()
     else:
         Migrate.diff_models(old_models_describe, models_describe)
         Migrate.diff_models(models_describe, old_models_describe, False)
-    Migrate._merge_operators()
+        Migrate._merge_operators()
     if isinstance(Migrate.ddl, MysqlDDL):
         expected_upgrade_operators = {
             "ALTER TABLE `category` MODIFY COLUMN `name` VARCHAR(200)",
             "ALTER TABLE `category` MODIFY COLUMN `slug` VARCHAR(100) NOT NULL",
+            "ALTER TABLE `category` DROP INDEX `title`",
+            "ALTER TABLE `config` DROP COLUMN `name`",
+            "ALTER TABLE `config` DROP INDEX `name`",
             "ALTER TABLE `config` ADD `user_id` INT NOT NULL  COMMENT 'User'",
             "ALTER TABLE `config` ADD CONSTRAINT `fk_config_user_17daa970` FOREIGN KEY (`user_id`) REFERENCES `user` (`id`) ON DELETE CASCADE",
             "ALTER TABLE `config` ALTER COLUMN `status` DROP DEFAULT",
@@ -812,6 +880,8 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `email` ADD `address` VARCHAR(200) NOT NULL",
             "ALTER TABLE `email` DROP COLUMN `user_id`",
             "ALTER TABLE `configs` RENAME TO `config`",
+            "ALTER TABLE `product` DROP COLUMN `uuid`",
+            "ALTER TABLE `product` DROP INDEX `uuid`",
             "ALTER TABLE `product` RENAME COLUMN `image` TO `pic`",
             "ALTER TABLE `email` RENAME COLUMN `id` TO `email_id`",
             "ALTER TABLE `product` ADD INDEX `idx_product_name_869427` (`name`, `type_db_alias`)",
@@ -827,16 +897,20 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `user` MODIFY COLUMN `is_active` BOOL NOT NULL  COMMENT 'Is Active' DEFAULT 1",
             "ALTER TABLE `user` MODIFY COLUMN `is_superuser` BOOL NOT NULL  COMMENT 'Is SuperUser' DEFAULT 0",
             "ALTER TABLE `user` MODIFY COLUMN `longitude` DECIMAL(10,8) NOT NULL",
-            "ALTER TABLE `user` ADD UNIQUE INDEX `uid_user_usernam_9987ab` (`username`)",
+            "ALTER TABLE `user` ADD UNIQUE INDEX `username` (`username`)",
             "CREATE TABLE `email_user` (\n    `email_id` INT NOT NULL REFERENCES `email` (`email_id`) ON DELETE CASCADE,\n    `user_id` INT NOT NULL REFERENCES `user` (`id`) ON DELETE CASCADE\n) CHARACTER SET utf8mb4",
             "CREATE TABLE IF NOT EXISTS `newmodel` (\n    `id` INT NOT NULL PRIMARY KEY AUTO_INCREMENT,\n    `name` VARCHAR(50) NOT NULL\n) CHARACTER SET utf8mb4",
             "ALTER TABLE `category` MODIFY COLUMN `created_at` DATETIME(6) NOT NULL  DEFAULT CURRENT_TIMESTAMP(6)",
             "ALTER TABLE `product` MODIFY COLUMN `body` LONGTEXT NOT NULL",
             "ALTER TABLE `email` MODIFY COLUMN `is_primary` BOOL NOT NULL  DEFAULT 0",
+            "CREATE TABLE `product_user` (\n    `product_id` INT NOT NULL REFERENCES `product` (`id`) ON DELETE CASCADE,\n    `user_id` INT NOT NULL REFERENCES `user` (`id`) ON DELETE CASCADE\n) CHARACTER SET utf8mb4",
         }
         expected_downgrade_operators = {
             "ALTER TABLE `category` MODIFY COLUMN `name` VARCHAR(200) NOT NULL",
             "ALTER TABLE `category` MODIFY COLUMN `slug` VARCHAR(200) NOT NULL",
+            "ALTER TABLE `category` ADD UNIQUE INDEX `title` (`title`)",
+            "ALTER TABLE `config` ADD `name` VARCHAR(100) NOT NULL UNIQUE",
+            "ALTER TABLE `config` ADD UNIQUE INDEX `name` (`name`)",
             "ALTER TABLE `config` DROP COLUMN `user_id`",
             "ALTER TABLE `config` DROP FOREIGN KEY `fk_config_user_17daa970`",
             "ALTER TABLE `config` ALTER COLUMN `status` SET DEFAULT 1",
@@ -845,15 +919,18 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `config` RENAME TO `configs`",
             "ALTER TABLE `product` RENAME COLUMN `pic` TO `image`",
             "ALTER TABLE `email` RENAME COLUMN `email_id` TO `id`",
+            "ALTER TABLE `product` ADD `uuid` INT NOT NULL UNIQUE",
+            "ALTER TABLE `product` ADD UNIQUE INDEX `uuid` (`uuid`)",
             "ALTER TABLE `product` DROP INDEX `idx_product_name_869427`",
             "ALTER TABLE `email` DROP INDEX `idx_email_email_4a1a33`",
             "ALTER TABLE `product` DROP INDEX `uid_product_name_869427`",
             "ALTER TABLE `product` ALTER COLUMN `view_num` DROP DEFAULT",
             "ALTER TABLE `user` ADD `avatar` VARCHAR(200) NOT NULL  DEFAULT ''",
-            "ALTER TABLE `user` DROP INDEX `idx_user_usernam_9987ab`",
+            "ALTER TABLE `user` DROP INDEX `username`",
             "ALTER TABLE `user` MODIFY COLUMN `password` VARCHAR(200) NOT NULL",
             "DROP TABLE IF EXISTS `email_user`",
             "DROP TABLE IF EXISTS `newmodel`",
+            "DROP TABLE IF EXISTS `product_user`",
             "ALTER TABLE `user` MODIFY COLUMN `intro` LONGTEXT NOT NULL",
             "ALTER TABLE `config` MODIFY COLUMN `value` TEXT NOT NULL",
             "ALTER TABLE `category` MODIFY COLUMN `created_at` DATETIME(6) NOT NULL  DEFAULT CURRENT_TIMESTAMP(6)",
@@ -866,6 +943,10 @@ def test_migrate(mocker: MockerFixture):
             "ALTER TABLE `product` MODIFY COLUMN `body` LONGTEXT NOT NULL",
             "ALTER TABLE `email` MODIFY COLUMN `is_primary` BOOL NOT NULL  DEFAULT 0",
         }
+        if should_add_user_id_column_type_alter_sql():
+            sql = "ALTER TABLE `category` MODIFY COLUMN `user_id` INT NOT NULL  COMMENT 'User'"
+            expected_upgrade_operators.add(sql)
+            expected_downgrade_operators.add(sql)
         assert not set(Migrate.upgrade_operators).symmetric_difference(expected_upgrade_operators)
 
         assert not set(Migrate.downgrade_operators).symmetric_difference(
@@ -874,9 +955,12 @@ def test_migrate(mocker: MockerFixture):
 
     elif isinstance(Migrate.ddl, PostgresDDL):
         expected_upgrade_operators = {
+            'DROP INDEX "uid_category_title_f7fc03"',
             'ALTER TABLE "category" ALTER COLUMN "name" DROP NOT NULL',
             'ALTER TABLE "category" ALTER COLUMN "slug" TYPE VARCHAR(100) USING "slug"::VARCHAR(100)',
             'ALTER TABLE "category" ALTER COLUMN "created_at" TYPE TIMESTAMPTZ USING "created_at"::TIMESTAMPTZ',
+            'ALTER TABLE "config" DROP COLUMN "name"',
+            'DROP INDEX "uid_config_name_2c83c8"',
             'ALTER TABLE "config" ADD "user_id" INT NOT NULL',
             'ALTER TABLE "config" ADD CONSTRAINT "fk_config_user_17daa970" FOREIGN KEY ("user_id") REFERENCES "user" ("id") ON DELETE CASCADE',
             'ALTER TABLE "config" ALTER COLUMN "status" DROP DEFAULT',
@@ -886,6 +970,8 @@ def test_migrate(mocker: MockerFixture):
             'ALTER TABLE "email" DROP COLUMN "user_id"',
             'ALTER TABLE "email" RENAME COLUMN "id" TO "email_id"',
             'ALTER TABLE "email" ALTER COLUMN "is_primary" TYPE BOOL USING "is_primary"::BOOL',
+            'DROP INDEX "uid_product_uuid_d33c18"',
+            'ALTER TABLE "product" DROP COLUMN "uuid"',
             'ALTER TABLE "product" ALTER COLUMN "view_num" SET DEFAULT 0',
             'ALTER TABLE "product" RENAME COLUMN "image" TO "pic"',
             'ALTER TABLE "product" ALTER COLUMN "is_reviewed" TYPE BOOL USING "is_reviewed"::BOOL',
@@ -904,11 +990,15 @@ def test_migrate(mocker: MockerFixture):
             'CREATE TABLE IF NOT EXISTS "newmodel" (\n    "id" SERIAL NOT NULL PRIMARY KEY,\n    "name" VARCHAR(50) NOT NULL\n);\nCOMMENT ON COLUMN "config"."user_id" IS \'User\'',
             'CREATE UNIQUE INDEX "uid_product_name_869427" ON "product" ("name", "type_db_alias")',
             'CREATE UNIQUE INDEX "uid_user_usernam_9987ab" ON "user" ("username")',
+            'CREATE TABLE "product_user" (\n    "product_id" INT NOT NULL REFERENCES "product" ("id") ON DELETE CASCADE,\n    "user_id" INT NOT NULL REFERENCES "user" ("id") ON DELETE CASCADE\n)',
         }
         expected_downgrade_operators = {
+            'CREATE UNIQUE INDEX "uid_category_title_f7fc03" ON "category" ("title")',
             'ALTER TABLE "category" ALTER COLUMN "name" SET NOT NULL',
             'ALTER TABLE "category" ALTER COLUMN "slug" TYPE VARCHAR(200) USING "slug"::VARCHAR(200)',
             'ALTER TABLE "category" ALTER COLUMN "created_at" TYPE TIMESTAMPTZ USING "created_at"::TIMESTAMPTZ',
+            'ALTER TABLE "config" ADD "name" VARCHAR(100) NOT NULL UNIQUE',
+            'CREATE UNIQUE INDEX "uid_config_name_2c83c8" ON "config" ("name")',
             'ALTER TABLE "config" ALTER COLUMN "status" SET DEFAULT 1',
             'ALTER TABLE "config" DROP COLUMN "user_id"',
             'ALTER TABLE "config" DROP CONSTRAINT "fk_config_user_17daa970"',
@@ -918,6 +1008,8 @@ def test_migrate(mocker: MockerFixture):
             'ALTER TABLE "email" DROP COLUMN "address"',
             'ALTER TABLE "email" RENAME COLUMN "email_id" TO "id"',
             'ALTER TABLE "email" ALTER COLUMN "is_primary" TYPE BOOL USING "is_primary"::BOOL',
+            'ALTER TABLE "product" ADD "uuid" INT NOT NULL UNIQUE',
+            'CREATE UNIQUE INDEX "uid_product_uuid_d33c18" ON "product" ("uuid")',
             'ALTER TABLE "product" ALTER COLUMN "view_num" DROP DEFAULT',
             'ALTER TABLE "product" RENAME COLUMN "pic" TO "image"',
             'ALTER TABLE "user" ADD "avatar" VARCHAR(200) NOT NULL  DEFAULT \'\'',
@@ -930,13 +1022,18 @@ def test_migrate(mocker: MockerFixture):
             'ALTER TABLE "product" ALTER COLUMN "created_at" TYPE TIMESTAMPTZ USING "created_at"::TIMESTAMPTZ',
             'ALTER TABLE "product" ALTER COLUMN "is_reviewed" TYPE BOOL USING "is_reviewed"::BOOL',
             'ALTER TABLE "product" ALTER COLUMN "body" TYPE TEXT USING "body"::TEXT',
+            'DROP TABLE IF EXISTS "product_user"',
             'DROP INDEX "idx_product_name_869427"',
             'DROP INDEX "idx_email_email_4a1a33"',
-            'DROP INDEX "idx_user_usernam_9987ab"',
+            'DROP INDEX "uid_user_usernam_9987ab"',
             'DROP INDEX "uid_product_name_869427"',
             'DROP TABLE IF EXISTS "email_user"',
             'DROP TABLE IF EXISTS "newmodel"',
         }
+        if should_add_user_id_column_type_alter_sql():
+            sql = 'ALTER TABLE "category" ALTER COLUMN "user_id" TYPE INT USING "user_id"::INT'
+            expected_upgrade_operators.add(sql)
+            expected_downgrade_operators.add(sql)
         assert not set(Migrate.upgrade_operators).symmetric_difference(expected_upgrade_operators)
         assert not set(Migrate.downgrade_operators).symmetric_difference(
             expected_downgrade_operators
@@ -966,3 +1063,15 @@ def test_sort_all_version_files(mocker):
         "10_datetime_update.py",
         "11_datetime_update.py",
     ]
+
+
+async def test_empty_migration(mocker, tmp_path: Path) -> None:
+    mocker.patch("os.listdir", return_value=[])
+    Migrate.app = "foo"
+    expected_content = MIGRATE_TEMPLATE.format(upgrade_sql="", downgrade_sql="")
+    Migrate.migrate_location = tmp_path
+
+    migration_file = await Migrate.migrate("update", True)
+
+    f = tmp_path / migration_file
+    assert f.read_text() == expected_content